stock data extraction into "cp" variable, refactor of part-2 code
Each time the program runs, a random stock is chosen from an array of
20 different stocks.  A call is made to the google finance api, and the
cp (change in percentage) value is returned. A negative cp indicates a
decrease in value, while a positive cp indicates an increase in value.
This method could be expanded to fetch other parameters of data, and
could also be altered to take user input.